requirements.txt
selenium
python-decouple
pdfplumber
